Seperated redux field normalizer functions to own file
